requirements.txt
gymnasium>=0.29.0
stable-baselines3>=2.1.0
tensorboard>=2.14.0
yfinance>=0.2.28
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
pyyaml>=6.0
datasets>=2.14.0
requests>=2.28.0
